Confirmation de la suppression avec l'action du snakebar

+++ app/src/main/java/com/megamind/todoapp/ui/screens/DetailScreen.kt
@@ -23,6 +23,11 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SnackbarDuration
+import androidx.compose.material3.SnackbarHost
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.material3.SnackbarResult
+
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
@@ -31,6 +36,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -45,6 +51,7 @@ import com.megamind.todoapp.R
 import com.megamind.todoapp.models.Task
 import com.megamind.todoapp.ui.theme.ToDoAppTheme
 import com.megamind.todoapp.viewModels.TaskViewModel
+import kotlinx.coroutines.launch
 import java.util.Date
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -57,7 +64,8 @@ fun DetailScreen(
 ) {
 
     val task by viewModel.fetchTaskById(id.toInt()).observeAsState()
-   //val snackbarHostState = rememberSnackbarHostState()
+    val snackbarHostState = remember { SnackbarHostState() }
+    val coroutineScope = rememberCoroutineScope()
 
     Surface {
         Scaffold(
@@ -80,7 +88,7 @@ fun DetailScreen(
                         }
                     })
             },
-
+            snackbarHost = { SnackbarHost(snackbarHostState) },
             content = {
 
                 Column(
@@ -90,13 +98,37 @@ fun DetailScreen(
                 ) {
 
                     task?.let { DetailBox(task = it) }
-                    Spacer(modifier = Modifier
-                        .height(50.dp)
-                        .padding(it))
+                    Spacer(
+                        modifier = Modifier
+                            .height(50.dp)
+                            .padding(it)
+                    )
                     ActionButtons(
                         onDelete = {
-                            task?.let { viewModel.deleteTask(it) }
-                            navController.popBackStack()
+
+                            coroutineScope.launch {
+                                val deletePremission = snackbarHostState.showSnackbar(
+                                    message = "Voulez vous vraiment supprimer?",
+                                    actionLabel = "Oui",
+                                    duration = SnackbarDuration.Long
+
+                                )
+
+                                when (deletePremission) {
+                                    SnackbarResult.Dismissed -> {
+                                    }
+
+                                    SnackbarResult.ActionPerformed -> {
+                                        task?.let { viewModel.deleteTask(it) }
+                                        navController.popBackStack()
+                                    }
+
+                                }
+
+
+                            }
+
+
                         },
                         onEdit = {
